requirements.txt
playwright==1.42.0
python-dotenv==1.0.1
click==8.1.7
beautifulsoup4==4.12.3
urllib3==2.2.1 